move extensions folder
