+++ stac_fastapi/elasticsearch/tests/conftest.py
@@ -15,6 +15,7 @@ from stac_fastapi.core.core import (
     CoreClient,
     TransactionsClient,
 )
+from stac_fastapi.core.extensions import QueryExtension
 from stac_fastapi.elasticsearch.config import (
     AsyncElasticsearchSettings,
     ElasticsearchSettings,
@@ -23,7 +24,6 @@ from stac_fastapi.elasticsearch.database_logic import (
     DatabaseLogic,
     create_collection_index,
 )
-from stac_fastapi.elasticsearch.extensions import QueryExtension
 from stac_fastapi.extensions.core import (  # FieldsExtension,
     ContextExtension,
     FieldsExtension,
